[#219] refactor ConversionService, CreateTablesService, ExporterService
