+++ lib/register_ingester_oc/companies/exporter_service.rb
@@ -2,10 +2,11 @@
 
 require_relative '../config/adapters'
 require_relative '../config/settings'
+require_relative '../services/exporter_service'
 
 module RegisterIngesterOc
   module Companies
-    class ExporterService
+    class ExporterService < Services::ExporterService
       # rubocop:disable Metrics/ParameterLists
       def initialize(
         athena_adapter: Config::Adapters::ATHENA_ADAPTER,
@@ -15,12 +16,8 @@ module RegisterIngesterOc
         full_s3_prefix: ENV.fetch('COMPANIES_EXPORT_JSON_FULL_S3_PREFIX'),
         diffs_s3_prefix: ENV.fetch('COMPANIES_EXPORT_JSON_DIFFS_S3_PREFIX')
       )
-        @athena_adapter = athena_adapter
-        @athena_database = athena_database
-        @s3_bucket = s3_bucket
-        @output_location = "s3://#{s3_bucket}/athena_results"
-        @filtered_table_name = filtered_table_name
-        @full_s3_prefix = full_s3_prefix
+        super(athena_adapter:, athena_database:, s3_bucket:,
+              filtered_table_name:, full_s3_prefix:)
         @diffs_s3_prefix = diffs_s3_prefix
       end
       # rubocop:enable Metrics/ParameterLists
@@ -32,43 +29,25 @@ module RegisterIngesterOc
         export_json_diffs month
       end
 
-      def s3_export_location_full(month)
-        "#{File.join("s3://#{s3_bucket}", full_s3_prefix, "mth=#{month}")}/"
-      end
-
       def s3_export_location_diffs(month)
         "#{File.join("s3://#{s3_bucket}", diffs_s3_prefix, "mth=#{month}")}/"
       end
 
       private
 
-      attr_reader :athena_adapter, :athena_database, :s3_bucket, :output_location, :filtered_table_name,
-                  :full_s3_prefix, :diffs_s3_prefix
+      attr_reader :diffs_s3_prefix
 
-      def export_all_json(month)
-        dst_table_name = "oc_export_full_#{month}"
-        dst_s3_location = s3_export_location_full(month)
+      def calc_prev_month(month)
+        month_split = month.split('_', 2).map(&:to_i)
+        raise 'wrong month format' if month_split.length != 2
 
-        query = <<~SQL
-          DROP TABLE IF EXISTS #{dst_table_name}
-        SQL
-        execute_sql query
+        year_i = month_split[0].to_i
+        month_i = month_split[1].to_i
 
-        query = <<~SQL
-          CREATE TABLE #{dst_table_name}
-          WITH (
-            format='JSON',
-            write_compression='GZIP',
-            external_location = '#{dst_s3_location}'
-          ) AS
-          SELECT
-            *
-          FROM
-            #{filtered_table_name} tab
-          WHERE
-            tab.mth = '#{month}';
-        SQL
-        execute_sql query
+        prev_year_i = month_i == 1 ? (year_i - 1) : year_i
+        prev_month_i = month_i == 1 ? 12 : (month_i - 1)
+
+        format('%d_%02d', prev_year_i, prev_month_i) # rubocop:disable Style/FormatStringToken
       end
 
       def export_json_diffs(month)
@@ -115,31 +94,6 @@ module RegisterIngesterOc
           );
         SQL
         execute_sql query
-      end
-
-      def execute_sql(sql_query)
-        athena_query = athena_adapter.start_query_execution(
-          {
-            query_string: sql_query,
-            result_configuration: {
-              output_location:
-            }
-          }
-        )
-        athena_adapter.wait_for_query(athena_query.query_execution_id)
-      end
-
-      def calc_prev_month(month)
-        month_split = month.split('_', 2).map(&:to_i)
-        raise 'wrong month format' if month_split.length != 2
-
-        year_i = month_split[0].to_i
-        month_i = month_split[1].to_i
-
-        prev_year_i = month_i == 1 ? (year_i - 1) : year_i
-        prev_month_i = month_i == 1 ? 12 : (month_i - 1)
-
-        format('%d_%02d', prev_year_i, prev_month_i) # rubocop:disable Style/FormatStringToken
       end
     end
   end
